feat(cubejs-cli): Move token command to TS
